Refactored component and organised it into folder.
